scripts/requirements.txt
click
imagehash
Pillow
